[LegacyTabs|LegacyFilters] Add legacy versions of Tabs and Filters (#8682)
### WHY are these changes introduced?

As part of the IndexFilters rollout plan, we want to separate out the
creation of LegacyTabs and LegacyFilters so that we can create PRs with
more focus and less mess. Creating this PR in this staggered way also
allows us to run a Polaris migration on Shopify/web to automatically
update all instances of Tabs and Filters to LegacyTabs and
LegacyFilters.

### WHAT is this pull request doing?

This PR is copying the Tabs and Filters components wholesale to
LegacyTabs and LegacyFilters. We also add documentation pages for these
new components.

### 🎩 checklist

- [x] Tested on
[mobile](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md#cross-browser-testing)
- [x] Tested on [multiple
browsers](https://help.shopify.com/en/manual/shopify-admin/supported-browsers)
- [x] Tested for
[accessibility](https://github.com/Shopify/polaris/blob/main/documentation/Accessibility%20testing.md)
- [x] Updated the component's `README.md` with documentation changes
- [x] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide