[IndexFilters] Support sort-only mode (#8876)
### WHY are these changes introduced?

We want to support the use case of not having any searching or filtering
functionality, but still retaining the Tabs and the sorting
functionality. We already support both the `hideQueryField` and
`hideFilters` props, but hadn't correctly handled the use-case when both
are true. We now will remove the toggle button to switch to the filter
mode, so that the UI is only the Tabs and the Sort button.

Whilst I was working around this component, I also fixed a bug around
the disabling of disabled filter pills. The desired effect is to render
the component, but return `null` from the component, rather than render
nothing at all. This prevents a bug where if a filter was previously
disabled and then becomes enabled, the Popover would automatically open
without any user interaction.

### 🎩 checklist

- [x] Tested on
[mobile](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md#cross-browser-testing)
- [x] Tested on [multiple
browsers](https://help.shopify.com/en/manual/shopify-admin/supported-browsers)
- [x] Tested for
[accessibility](https://github.com/Shopify/polaris/blob/main/documentation/Accessibility%20testing.md)
- [x] Updated the component's `README.md` with documentation changes
- [x] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide