[Navigation] Add alternative icon for matched item (#9527)
### WHAT is this pull request doing?

Introduce a new prop called `matchedItemIcon` for the `Item` component
within the `Navigation`. This prop accepts an icon source that will be
displayed when the item is selected. Although the `icon` prop can be
used to toggle the icon by the consumer, in most cases the selected item
is determined by the `pathname` provided to the `Navigation`. As a
result, the decision to display the appropriate icon should be made by
the `Item` component itself.




https://github.com/Shopify/polaris/assets/2091116/f980f42d-81a0-4566-a7b4-73a8064faaee






[Figma](https://www.figma.com/file/jLLkmo9r28hiqPvf4s90E4/Polaris-Uplift-Components-%5Bgen3%E2%80%93alpha%5D?type=design&node-id=48-108&mode=design&t=wvAxGDVK27h6diwG-0)

### 🎩 checklist

- [x] Tested on
[mobile](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md#cross-browser-testing)
- [x] Tested on [multiple
browsers](https://help.shopify.com/en/manual/shopify-admin/supported-browsers)
- [x] Tested for
[accessibility](https://github.com/Shopify/polaris/blob/main/documentation/Accessibility%20testing.md)
- [ ] Updated the component's `README.md` with documentation changes
- [ ] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide

---------

Co-authored-by: Kyle Durand <kyledurand@users.noreply.github.com>