[polaris.shopify.com] Fix icon search (#9557)
I've been working quite a bit with icons recently and this has been
driving me crazy. Typing quickly in the icon search field often leads to
incorrect results.

This is me typing `add note`. It only registered half the characters


https://github.com/Shopify/polaris/assets/6844391/bc84d0c0-eac7-45dd-9b60-e16343f9d6f1

Simple solution: Maintain logic but it doesn't need to be a controlled
component